first, let's copy the changes from fruition's caching_query here.
- handle lock errors from redis and dogpile, by just running the original query.
- some minor py3 compatibility in generating the query key
- handle cache key prefixes

NOTE: I changed the behavior of the cache key prefix code.
What was there was nonsense, so I tried to interpret the intended behavior.

It was saying `if getattr(self, '_cache_region')` in a place where there was *definitely* a `self._cache_region` defined, so at the least it could have been written as `if self._cache_region`.
However, it then went on to use `self._cache_region.cache_prefix`, and I believe the intent was to check if `cache_prefix` was defined, so I changed it to `if getattr(self._cache_region, 'cache_prefix')`.